requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0

# LangChain Ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-mistralai>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# Vector Store & Search
faiss-cpu>=1.8.0
tavily-python>=0.5.0
langchain-tavily>=0.0.1

# Utilities
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
httpx>=0.27.0

# Logging
structlog>=24.0.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Serverless Adapter
mangum>=0.17.0
